Fix camera framerate type (#76)
* fix type of camera framerate setting

* fix CameraSettigns.framerate type

+++ clients/rust/src/camera_settings.rs
@@ -1,12 +1,12 @@
 // CameraSettings represents a CameraSettings model.
-#[derive(Clone, Debug, Deserialize, PartialEq, PartialOrd, Serialize)]
+#[derive(Clone, Debug, Deserialize, Eq, Hash, Ord, PartialEq, PartialOrd, Serialize)]
 pub struct CameraSettings {
     #[serde(rename="height")]
     pub height: i32,
     #[serde(rename="width")]
     pub width: i32,
     #[serde(rename="framerate")]
-    pub framerate: f64,
+    pub framerate: i32,
     #[serde(rename="device_name")]
     pub device_name: String,
     #[serde(rename="label")]
@@ -14,7 +14,7 @@ pub struct CameraSettings {
 }
 
 impl CameraSettings {
-    pub fn new(height: i32, width: i32, framerate: f64, device_name: String, label: String) -> CameraSettings {
+    pub fn new(height: i32, width: i32, framerate: i32, device_name: String, label: String) -> CameraSettings {
         CameraSettings {
             height,
             width,
